requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
--find-links https://data.dgl.ai/wheels/repo.html
spacy-universal-sentence-encoder
torch==1.7.0+cu110
dgl-cu110==0.5.3
networkx==2.2
numpy==1.21.2
scikit-learn==0.24.2
scipy==1.7.1
pyemd
tensorboard
tqdm
git+https://github.com/fabriziocosta/EDeN.git
natsort
SPARQLWrapper==1.8.5
auto-sklearn==0.14.0
FLAML==0.6.3
mindware==0.5.0
pandas
matplotlib


# dataset embedding requirements
bitstring==3.1.9
swifter
